Added a warning constraint on a missing event binding to ECNO nets.
This warning indicates when a transition is missing an event binding
(either because it was not parsable or was completely forgotten). Fixed
a long overlooked problem, where the plugin of the generated package for
ECNO nets referred to an old genmodel (ECLANets.genmodel), which however
did not have any practical impact. Increased the version number of the
ECNO Net plugin to 0.4.1.